[구성] MainView 컴포넌트 분리
