+++ src/components/Home/index.tsx
@@ -1,0 +1,3 @@
+import MainView from './MainView';
+
+export { MainView };
